+++ App/Swift Co-Pilot/ObjectDetectionViewController.swift
@@ -207,27 +207,24 @@ extension ObjectDetectionViewController {
         for i in 0..<totalElements {
             switch elements[i] {
             case "Buttons":
-                swiftCodes.append("UIButton")
-                swiftUICode.append("SwiftUI Buttons")
+                swiftCodes.append(snippets.get.buttons[0])
+                swiftUICode.append(snippets.get.buttons[1])
             case "Image":
-                swiftCodes.append("UIImage")
-                swiftUICode.append("SwiftUI Image")
+                swiftCodes.append(snippets.get.image[0])
+                swiftUICode.append(snippets.get.image[1])
             case "TF":
-                swiftCodes.append("UITextField")
-                swiftUICode.append("SwiftUI TextField")
+                swiftCodes.append(snippets.get.textField[0])
+                swiftUICode.append(snippets.get.textField[1])
             default:
                 errorAlert(mesg: "No elements found")
             }
         }
-        
-        print(swiftCodes)
-        print(swiftUICode)
-        
+       
         ViewController.resultSnippet.removeAll()
         
         
-        finalSwiftCode = self.swiftCodes.joined(separator: "\n")
-        swiftUIFinalCode = self.swiftUICode.joined(separator: "\n")
+        finalSwiftCode = self.swiftCodes.joined(separator: "\n\n")
+        swiftUIFinalCode = self.swiftUICode.joined(separator: "\n\n")
         
         ViewController.resultSnippet.append(finalSwiftCode)
         ViewController.resultSnippet.append(swiftUIFinalCode)
